feat: modularize feature of filtering status of assignement
